Updates in code and configuration to allow multi-domain configurations.

+++ xmlvalidator-web/src/main/java/eu/europa/ec/itb/einvoice/upload/FileController.java
@@ -1,33 +1,19 @@
 package eu.europa.ec.itb.einvoice.upload;
 
-import com.gitb.tr.ObjectFactory;
-import com.gitb.tr.TAR;
 import eu.europa.ec.itb.einvoice.ApplicationConfig;
+import eu.europa.ec.itb.einvoice.DomainConfig;
+import eu.europa.ec.itb.einvoice.DomainConfigCache;
 import eu.europa.ec.itb.einvoice.util.FileManager;
 import org.apache.commons.io.FileUtils;
-import org.apache.tika.Tika;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import org.slf4j.MDC;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.io.FileSystemResource;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.web.bind.annotation.*;
-import org.w3c.dom.Document;
 
 import javax.servlet.http.HttpServletResponse;
-import javax.xml.bind.JAXBContext;
-import javax.xml.bind.JAXBException;
-import javax.xml.bind.Marshaller;
-import javax.xml.parsers.DocumentBuilderFactory;
-import javax.xml.transform.OutputKeys;
-import javax.xml.transform.Transformer;
-import javax.xml.transform.TransformerFactory;
-import javax.xml.transform.dom.DOMSource;
-import javax.xml.transform.stream.StreamResult;
-import java.io.*;
-import java.util.UUID;
+import java.io.File;
 
 /**
  * Created by simatosc on 08/03/2016.
@@ -35,29 +21,39 @@ import java.util.UUID;
 @RestController
 public class FileController {
 
-    private static Logger logger = LoggerFactory.getLogger(FileController.class);
-
     @Autowired
     ApplicationConfig config;
     @Autowired
     FileManager fileManager;
     @Autowired
     ReportGeneratorBean reportGenerator;
+    @Autowired
+    DomainConfigCache domainConfigCache;
 
-    @RequestMapping(value = "/xml/{id}", method = RequestMethod.GET, produces = MediaType.TEXT_PLAIN_VALUE)
+    @RequestMapping(value = "/{domain}/xml/{id}", method = RequestMethod.GET, produces = MediaType.TEXT_PLAIN_VALUE)
     @ResponseBody
-    public FileSystemResource getXML(@PathVariable String id) {
+    public FileSystemResource getXML(@PathVariable String domain, @PathVariable String id) {
+        DomainConfig domainConfig = domainConfigCache.getConfigForDomain(domain);
+        if (domainConfig == null) {
+            throw new NotFoundException();
+        }
+        MDC.put("domain", domain);
         File reportFile = new File(config.getReportFolder(), fileManager.getInputFileName(id));
         if (reportFile.exists() && reportFile.isFile()) {
             return new FileSystemResource(reportFile);
         } else {
-            throw new ResourceNotFoundException();
+            throw new NotFoundException();
         }
     }
 
-    @RequestMapping(value = "/report/{id}/xml", method = RequestMethod.GET, produces = MediaType.APPLICATION_OCTET_STREAM_VALUE)
+    @RequestMapping(value = "/{domain}/report/{id}/xml", method = RequestMethod.GET, produces = MediaType.TEXT_PLAIN_VALUE)
     @ResponseBody
-    public FileSystemResource getReportXml(@PathVariable String id, HttpServletResponse response) {
+    public FileSystemResource getReportXml(@PathVariable String domain, @PathVariable String id, HttpServletResponse response) {
+        DomainConfig domainConfig = domainConfigCache.getConfigForDomain(domain);
+        if (domainConfig == null) {
+            throw new NotFoundException();
+        }
+        MDC.put("domain", domain);
         File reportFile = new File(config.getReportFolder(), fileManager.getReportFileNameXml(id));
         if (reportFile.exists() && reportFile.isFile()) {
             if (response != null) {
@@ -65,21 +61,26 @@ public class FileController {
             }
             return new FileSystemResource(reportFile);
         } else {
-            throw new ResourceNotFoundException();
+            throw new NotFoundException();
         }
     }
 
-    @RequestMapping(value = "/report/{id}/pdf", method = RequestMethod.GET, produces = MediaType.APPLICATION_OCTET_STREAM_VALUE)
+    @RequestMapping(value = "/{domain}/report/{id}/pdf", method = RequestMethod.GET, produces = MediaType.APPLICATION_OCTET_STREAM_VALUE)
     @ResponseBody
-    public FileSystemResource getReportPdf(@PathVariable String id, HttpServletResponse response) {
+    public FileSystemResource getReportPdf(@PathVariable String domain, @PathVariable String id, HttpServletResponse response) {
+        DomainConfig domainConfig = domainConfigCache.getConfigForDomain(domain);
+        if (domainConfig == null) {
+            throw new NotFoundException();
+        }
+        MDC.put("domain", domain);
         File reportFile = new File(config.getReportFolder(), fileManager.getReportFileNamePdf(id));
         if (!(reportFile.exists() && reportFile.isFile())) {
             // Generate the PDF.
             File reportFileXml = new File(config.getReportFolder(), fileManager.getReportFileNameXml(id));
             if (reportFileXml.exists() && reportFileXml.isFile()) {
-                reportGenerator.writeReport(reportFileXml, reportFile);
+                reportGenerator.writeReport(domainConfig, reportFileXml, reportFile);
             } else {
-                throw new ResourceNotFoundException();
+                throw new NotFoundException();
             }
         }
         if (response != null) {
@@ -88,17 +89,22 @@ public class FileController {
         return new FileSystemResource(reportFile);
     }
 
-    public FileSystemResource getReportXml(String id) {
-        return this.getReportXml(id, null);
+    public FileSystemResource getReportXml(String domain, String id) {
+        return this.getReportXml(domain, id, null);
     }
 
-    public FileSystemResource getReportPdf(String id) {
-        return this.getReportPdf(id, null);
+    public FileSystemResource getReportPdf(String domain, String id) {
+        return this.getReportPdf(domain, id, null);
     }
 
-    @RequestMapping(value = "/report/{id}", method = RequestMethod.DELETE)
+    @RequestMapping(value = "/{domain}/report/{id}", method = RequestMethod.DELETE)
     @ResponseBody
-    public void deleteReport(@PathVariable String id) {
+    public void deleteReport(@PathVariable String domain, @PathVariable String id) {
+        DomainConfig domainConfig = domainConfigCache.getConfigForDomain(domain);
+        if (domainConfig == null) {
+            throw new NotFoundException();
+        }
+        MDC.put("domain", domain);
         File reportFile = new File(config.getReportFolder(), fileManager.getReportFileNameXml(id));
         if (reportFile.exists() && reportFile.isFile()) {
             FileUtils.deleteQuietly(reportFile);
@@ -109,9 +115,14 @@ public class FileController {
         }
     }
 
-    @RequestMapping(value = "/xml/{id}", method = RequestMethod.DELETE)
+    @RequestMapping(value = "/{domain}/xml/{id}", method = RequestMethod.DELETE)
     @ResponseBody
-    public void deleteXML(@PathVariable String id) {
+    public void deleteXML(@PathVariable String domain, @PathVariable String id) {
+        DomainConfig domainConfig = domainConfigCache.getConfigForDomain(domain);
+        if (domainConfig == null) {
+            throw new NotFoundException();
+        }
+        MDC.put("domain", domain);
         File reportFile = new File(config.getReportFolder(), fileManager.getInputFileName(id));
         if (reportFile.exists() && reportFile.isFile()) {
             FileUtils.deleteQuietly(reportFile);
@@ -143,10 +154,6 @@ public class FileController {
             }
         }
         return handled;
-    }
-
-    @ResponseStatus(value = HttpStatus.NOT_FOUND)
-    public static class ResourceNotFoundException extends RuntimeException {
     }
 
 }
